# Label optional dependencies as external