@50Wliu 2021-11-19
By the way, is the source code of `microsoft.aspnetcore.razor.vscode` public? I'd like to get rid of the vscode-nls warnings as well, but it looks like they're coming by way of that package and I can't find the repository for it.