@JoeRobich 2021-11-19
> is the source code of `microsoft.aspnetcore.razor.vscode` public?

I believe you can find it at https://github.com/dotnet/razor-tooling